fix(contracts): money and controlled-substance quantities were float (LEAD-009/010) (#212)
`payment_contracts.py` has stated the package convention since it was
written — "Monetary amounts are represented as `Decimal` in major
currency units" — and `payment_contracts.py` / `finance_contracts.py`
follow it. Eight other modules did not. 79 fields across them carried
money or a counted controlled substance as `float`.

IEEE-754 binary cannot represent 0.1 or 2.675. A waste amount a
clinician enters as 0.1 mg is stored as
0.1000000000000000055511151231257827, and a sum of per-administration
waste does not reconcile against the recorded doses. On a DEA-reportable
ledger an unexplained fractional discrepancy is not a rounding annoyance
— it reads as diversion, and the record cannot be defended because the
arithmetic itself is lossy.

## Scope, and where it differs from the register

LEAD-009 names five fields in `narcotic.py`. **There are 24.**
`quantity`, `dose`, `pre_count`, `post_count`, `expected_count`,
`actual_count`, `variance` and `medication_counts` are the values waste
is computed *from*.

Converting only the five named `waste_amount` fields would have been
worse than leaving them alone: `Decimal - float` raises `TypeError` in
Python, so a half-converted ledger turns a wrong reconciliation into a
crashing one. All of `narcotic.py` moved together. Nothing in the module
does arithmetic or has a validator, so the conversion is annotation-only
there.

LEAD-010 places `inventory_events.py`, `medications_events.py`,
`narcotics_events.py` and `supply_integrations.py` under `schemas/`;
they live at the package root. Its line references are otherwise
accurate.

| Module | Fields |
|---|---|
| `schemas/narcotic.py` | 24 |
| `schemas/inventory_contracts.py` | 21 |
| `schemas/founder_contracts.py` | 7 |
| `schemas/crm_contracts.py` | 8 |
| `inventory_events.py` | 8 |
| `medications_events.py` | 8 |
| `narcotics_events.py` | 3 |
| `supply_integrations.py` | 3 (function parameters) |

Deliberately left as `float`: `churn_rate`, `net_retention_rate`,
`denial_rate`, `appeal_success_rate`, `average_days_to_payment`,
`variance_pct`, `funding_percentage`, `likelihood_percentage`,
`readiness_score`, `risk_score` and the trend lists. Rates, percentages,
scores and durations are neither money nor a counted quantity, and
converting them would break their consumers for nothing.

`gt=0` and `ge=0` were preserved per field rather than normalised —
`WasteNarcoticRequest.waste_amount` still refuses zero,
`MedicationCreateRequest.quantity` still accepts it. Both are asserted.

## The breaking part, stated plainly

Pydantic serialises `Decimal` to a JSON **string**. Every one of these
79 fields previously emitted a JSON number; they now emit `"2.675"`.
Input is unaffected — a JSON number still deserialises — so an
unmigrated producer keeps working against a migrated consumer. A
consumer that does arithmetic on the parsed value, or validates `type:
number`, breaks.

That is already the behaviour of the package's own `Decimal` fields
(`InvoiceRef.amount_due` emits a string today), so this makes the
package self-consistent rather than inventing a third convention.

## No version is cut

`pyproject.toml` stays at 2.36.0 on purpose. Cutting a major release
now, with zero consumers migrated and LEAD-019's convergence not done,
strands 63 repositories one major behind on a package whose gateway
audience pin is fail-closed in production. The release belongs with the
consumer wave, not ahead of it. `CHANGELOG.md` records this under
`[Unreleased]`.

## Verification

- `tests/test_money_and_quantity_precision.py`, 17 tests: 0.1, 2.675,
0.05, 1.005 and 0.3 each survive `model_dump_json()` →
`model_validate_json()` with their exact digits; the JSON member is
asserted to be a string so the wire break is pinned rather than
rediscovered; a JSON number is asserted to still deserialise; `gt` and
`ge` are asserted separately; and a regression guard fails if any field
matching
`(amount|cost|price|total|revenue|balance|paid|pledged|funded|committed|quantity|dose|waste|_count|counts|variance)`
is annotated `float` in the eight modules.
- That guard caught four fields the conversion missed on the first pass
— `medication_counts: dict[str, float]` and three `cost:
Optional[float]` — which is the reason it exists.
- Full package suite: **1769 passed, 1 xfailed**.
- `ruff format --check` clean on all nine changed files. `ruff check`
adds no new finding class beyond `FURB157` on `Decimal("0")`, which is
the exact spelling `payment_contracts.py` and `finance_contracts.py`
already use.

## Consumers still to migrate — none touched here

Confirmed by org code search before the GitHub search API rate-limited:

| Repo | Imports |
|---|---|
| `Adaptix-Narcotics-Service` | `schemas.narcotic`,
`supply_integrations` |
| `Adaptix-Medications-Service` | `medications_events`,
`supply_integrations` |
| `Adaptix-Inventory-Service` | `supply_integrations` |
| `Adaptix-CAD-Service` | `supply_integrations` |

**Not yet resolved:** consumers of `schemas/inventory_contracts.py`,
`schemas/founder_contracts.py` and `schemas/crm_contracts.py`. The
search for those three returned rate-limit errors rather than empty
results — treat the absence as unknown, not as "no consumers". Re-run:

```
gh api -X GET search/code -f q='"inventory_contracts import" org:FusionEMS-Quantum-LLC'
gh api -X GET search/code -f q='"founder_contracts import" org:FusionEMS-Quantum-LLC'
gh api -X GET search/code -f q='"crm_contracts import" org:FusionEMS-Quantum-LLC'
```

Each migrated consumer needs: `Decimal` on any local model mirroring
these fields; a check that no code path mixes the value with a `float`;
a DB column type check where the value is persisted (a `Float`/`REAL`
column re-introduces the defect after the contract fixed it, and
changing it needs a migration in that repo); and, for any TypeScript
consumer, `JSON.parse` now yielding a string where it yielded a number.

+++ adaptix_contracts/inventory_events.py
@@ -13,6 +13,7 @@ cross-service correlation and audit.
 from __future__ import annotations
 
 from datetime import datetime
+from decimal import Decimal
 from enum import Enum
 from typing import Optional, Any
 from uuid import UUID
@@ -56,7 +57,7 @@ class InventoryItemEvent(BaseModel):
     par_level: int = Field(..., description="Par stock level")
     current_stock: int = Field(..., description="Current stock count")
     unit: str = Field(..., description="Unit of measure")
-    cost_per_unit: float = Field(..., description="Cost per unit")
+    cost_per_unit: Decimal = Field(..., description="Cost per unit")
 
     # Before/after for updates
     before_state: Optional[dict[str, Any]] = Field(
@@ -91,8 +92,8 @@ class InventoryStockAdjustmentEvent(BaseModel):
     adjustment_reason: str = Field(..., description="Reason for adjustment")
 
     par_level: int = Field(..., description="Par stock level")
-    cost_per_unit: float = Field(..., description="Cost per unit")
-    adjustment_cost: float = Field(..., description="Cost impact of adjustment")
+    cost_per_unit: Decimal = Field(..., description="Cost per unit")
+    adjustment_cost: Decimal = Field(..., description="Cost impact of adjustment")
 
     actor_user_id: Optional[str] = Field(None, description="User performing action")
     timestamp: datetime = Field(..., description="Event timestamp")
@@ -118,8 +119,8 @@ class InventoryLowStockAlert(BaseModel):
         ..., description="Recommended reorder amount"
     )
 
-    cost_per_unit: float = Field(..., description="Cost per unit")
-    total_reorder_cost: float = Field(
+    cost_per_unit: Decimal = Field(..., description="Cost per unit")
+    total_reorder_cost: Decimal = Field(
         ..., description="Est. cost of recommended reorder"
     )
 
@@ -149,8 +150,8 @@ class InventoryExpirationAlert(BaseModel):
     expiration_date: datetime = Field(..., description="Item expiration date")
     days_until_expiration: int = Field(..., description="Days until expiration")
     current_stock: int = Field(..., description="Stock count")
-    cost_per_unit: float = Field(..., description="Cost per unit")
-    waste_forecast: float = Field(..., description="Est. cost of waste if not used")
+    cost_per_unit: Decimal = Field(..., description="Cost per unit")
+    waste_forecast: Decimal = Field(..., description="Est. cost of waste if not used")
 
     notify_role: str = Field(default="supply_officer", description="Role to notify")
     severity: str = Field(
@@ -201,7 +202,7 @@ class InventoryAnalyticsEvent(BaseModel):
 
     # Generic metrics
     quantity: Optional[int] = Field(None, description="Quantity involved")
-    cost: Optional[float] = Field(None, description="Cost impact")
+    cost: Optional[Decimal] = Field(None, description="Cost impact")
 
     # Custom fields for different event types
     metadata: Optional[dict[str, Any]] = Field(
